feat: separated proxy and auth

+++ app/routers/impl/camera_router.py
@@ -1,9 +1,12 @@
+import io
 from typing import Sequence
 
 from fastapi import Request
 from fastapi.responses import StreamingResponse
 
+from app.clients.auth_client import AuthClient
 from app.config.bindings import inject
+from app.exceptions.authorization_exception import AuthorizationException
 from app.models.camera import Camera, CameraInputDto
 from app.routers.router_wrapper import RouterWrapper
 from app.services.camera.camera_service import CameraService
@@ -14,9 +17,10 @@ stream_manager = FFmpegStreamManager()
 
 class CameraRouter(RouterWrapper):
     @inject
-    def __init__(self, camera_service: CameraService):
+    def __init__(self, camera_service: CameraService, auth_client: AuthClient):
         super().__init__(prefix=f"/camera")
         self.camera_service = camera_service
+        self.auth_client = auth_client
 
     def _define_routes(self):
         @self.router.get("/{ip}")
@@ -25,11 +29,19 @@ class CameraRouter(RouterWrapper):
 
         @self.router.post("/", operation_id="create_camera_slash")
         @self.router.post("", operation_id="create_camera_without_slash")
-        def create_camera(camera: CameraInputDto) -> Camera:
+        async def create_camera(request: Request, camera: CameraInputDto) -> Camera:
+            token = request.headers.get("Authorization")
+            user = await self.auth_client.get_authenticated_user(token)
+            if user is None or "MODIFY_DEVICES" not in user.permissions:
+                raise AuthorizationException("Not authorized")
             return self.camera_service.create(Camera.from_dto(camera))
 
         @self.router.delete("/{ip}")
-        def delete_camera_by_ip(ip: str) -> Camera:
+        async def delete_camera_by_ip(request: Request, ip: str) -> Camera:
+            token = request.headers.get("Authorization")
+            user = await self.auth_client.get_authenticated_user(token)
+            if user is None or "MODIFY_DEVICES" not in user.permissions:
+                raise AuthorizationException("Not authorized")
             stream_manager.cleanup_stream(ip)
             return self.camera_service.delete_by_ip(ip)
 
@@ -39,6 +51,13 @@ class CameraRouter(RouterWrapper):
 
         @self.router.get("/{ip}/stream")
         async def get_camera_stream_by_ip(request: Request, ip: str):
+            token = request.headers.get("Authorization")
+            if token is None and request.query_params.get("auth_token") is not None:
+                token = "Bearer " + request.query_params.get("auth_token")
+            user = await self.auth_client.get_authenticated_user(token)
+            if user is None or "MODIFY_DEVICES" not in user.permissions:
+                raise AuthorizationException("Not authorized")
+
             camera = self.camera_service.get_by_ip(ip)
 
             async def generate():
