+++ app/routers/impl/recording_router.py
@@ -1,8 +1,9 @@
 from typing import Sequence
 from fastapi import Request, Query
 
+from app.clients.auth_client import AuthClient
 from app.config.bindings import inject
-from app.exceptions.bad_request_exception import BadRequestException
+from app.exceptions.authorization_exception import AuthorizationException
 from app.models.enums.recording_type import RecordingType
 from app.models.recording import Recording
 from app.routers.router_wrapper import RouterWrapper
@@ -11,43 +12,65 @@ from app.services.recording.recording_service import RecordingService
 
 class RecordingRouter(RouterWrapper):
     @inject
-    def __init__(self, recording_service: RecordingService):
+    def __init__(self, recording_service: RecordingService, auth_client: AuthClient):
         super().__init__(prefix=f"/recording")
         self.recording_service = recording_service
-
+        self.auth_client = auth_client
 
     def _define_routes(self):
         # Basic CRUD
         @self.router.get("/{rec_id}")
-        def get_recording_by_id(rec_id: int) -> Recording:
+        async def get_recording_by_id(request: Request, rec_id: int) -> Recording:
+            token = request.headers.get("Authorization")
+            user = await self.auth_client.get_authenticated_user(token)
+            if user is None or "ACCESS_RECORDINGS" not in user.permissions:
+                raise AuthorizationException("Not authorized")
             return self.recording_service.get_by_id(rec_id)
 
-
         @self.router.delete("/{rec_id}")
-        def delete_recording_by_id(rec_id: int) -> Recording:
+        async def delete_recording_by_id(request: Request, rec_id: int) -> Recording:
+            token = request.headers.get("Authorization")
+            user = await self.auth_client.get_authenticated_user(token)
+            if user is None or "ACCESS_RECORDINGS" not in user.permissions:
+                raise AuthorizationException("Not authorized")
             return self.recording_service.delete_by_id(rec_id)
 
-
         @self.router.delete("/")
-        def delete_recording_by_id() -> Sequence[Recording]:
+        async def delete_recording_by_id(request: Request) -> Sequence[Recording]:
+            token = request.headers.get("Authorization")
+            user = await self.auth_client.get_authenticated_user(token)
+            if user is None or "ACCESS_RECORDINGS" not in user.permissions:
+                raise AuthorizationException("Not authorized")
             return self.recording_service.delete_all()
 
-
         @self.router.get("/{rec_id}/stream")
-        def stream_recording(request: Request, rec_id: int):
+        async def stream_recording(request: Request, rec_id: int):
+            token = request.headers.get("Authorization")
+            if token is None and request.query_params.get("auth_token") is not None:
+                token = "Bearer " + request.query_params.get("auth_token")
+            user = await self.auth_client.get_authenticated_user(token)
+            if user is None or "ACCESS_RECORDINGS" not in user.permissions:
+                raise AuthorizationException("Not authorized")
             return self.recording_service.stream(request, rec_id)
 
-
         @self.router.get("/{rec_id}/download")
-        def download_recording(rec_id: int):
+        async def download_recording(request: Request, rec_id: int):
+            token = request.headers.get("Authorization")
+            user = await self.auth_client.get_authenticated_user(token)
+            if user is None or "ACCESS_RECORDINGS" not in user.permissions:
+                raise AuthorizationException("Not authorized")
             return self.recording_service.download(rec_id)
-
 
         # Other endpoints
         @self.router.get("/", operation_id="get_all_recordings_with_slash")
         @self.router.get("", operation_id="get_all_recordings_without_slash")
-        def get_all_recordings(
+        async def get_all_recordings(
+                request: Request,
                 offset: int = 0,
                 rec_type: RecordingType | None = Query(default=None, alias="type")
         ) -> Sequence[Recording]:
+            token = request.headers.get("Authorization")
+            user = await self.auth_client.get_authenticated_user(token)
+            if user is None or "ACCESS_RECORDINGS" not in user.permissions:
+                raise AuthorizationException("Not authorized")
             return self.recording_service.get_all_paginated(offset=offset, recording_type=rec_type)
